# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ScholarForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/members:
    get:
      operationId: listProjectMembers
      tags: [members]
      summary: List project members
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectMember"
  /projects/{projectId}/members/{memberId}:
    patch:
      operationId: updateProjectMember
      tags: [members]
      summary: Update member role
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
        - in: path
          name: memberId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberBody"
      responses:
        "200":
          description: Member updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMember"
    delete:
      operationId: removeProjectMember
      tags: [members]
      summary: Remove member from project
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
        - in: path
          name: memberId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Member removed

  # Invitations
  /projects/{projectId}/invitations:
    post:
      operationId: createInvitation
      tags: [invitations]
      summary: Invite a collaborator by email
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationBody"
      responses:
        "201":
          description: Invitation sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
    get:
      operationId: listInvitations
      tags: [invitations]
      summary: List invitations for a project
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invitation"
  /invitations/{invitationId}/accept:
    post:
      operationId: acceptInvitation
      tags: [invitations]
      summary: Accept an invitation
      parameters:
        - in: path
          name: invitationId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invitation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /invitations/{invitationId}/decline:
    post:
      operationId: declineInvitation
      tags: [invitations]
      summary: Decline an invitation
      parameters:
        - in: path
          name: invitationId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invitation declined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /projects/{projectId}/milestones:
    get:
      operationId: listMilestones
      tags: [milestones]
      summary: List milestones for a project
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"
    post:
      operationId: createMilestone
      tags: [milestones]
      summary: Create a milestone
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMilestoneBody"
      responses:
        "201":
          description: Milestone created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
  /projects/{projectId}/milestones/{milestoneId}:
    patch:
      operationId: updateMilestone
      tags: [milestones]
      summary: Update a milestone
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
        - in: path
          name: milestoneId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMilestoneBody"
      responses:
        "200":
          description: Milestone updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
    delete:
      operationId: deleteMilestone
      tags: [milestones]
      summary: Delete a milestone
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
        - in: path
          name: milestoneId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Milestone deleted

  # Files
...
  /projects/{projectId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List chat messages for a project
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: before
          schema:
            type: string
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a chat message
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"

  # Activity Log
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        institution:
          type: ["string", "null"]
        researchInterests:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - createdAt
        - updatedAt
...
    ProjectSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        keywords:
          type: array
          items:
            type: string
        status:
          type: string
        visibility:
          type: string
        memberCount:
          type: integer
        taskCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        currentUserRole:
          type: ["string", "null"]
      required:
        - id
        - title
        - description
        - keywords
        - status
        - visibility
        - memberCount
        - taskCount
        - createdAt
        - updatedAt
...
    ProjectMember:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        projectId:
          type: string
        role:
          type: string
        joinedAt:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - id
        - userId
        - projectId
        - role
        - joinedAt
        - user
    UpdateMemberBody:
      type: object
      properties:
        role:
          type: string
      required:
        - role

    # Invitations
    Invitation:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        projectId:
          type: string
        invitedById:
          type: string
        status:
          type: string
        createdAt:
          type: string
        project:
          $ref: "#/components/schemas/ProjectSummary"
        invitedBy:
          $ref: "#/components/schemas/User"
      required:
        - id
        - email
        - projectId
        - invitedById
        - status
        - createdAt
        - project
        - invitedBy
    CreateInvitationBody:
      type: object
      properties:
        email:
          type: string
      required:
        - email

    # Tasks
...
    Milestone:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        title:
          type: string
        dueDate:
          type: string
        completed:
          type: boolean
      required:
        - id
        - projectId
        - title
        - dueDate
        - completed
    CreateMilestoneBody:
      type: object
      properties:
        title:
          type: string
        dueDate:
          type: string
      required:
        - title
        - dueDate
    UpdateMilestoneBody:
      type: object
      properties:
        title:
          type: string
        dueDate:
          type: string
        completed:
          type: boolean

    # Files
...
    ChatMessage:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        userId:
          type: string
        message:
          type: string
        createdAt:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - id
        - projectId
        - userId
        - message
        - createdAt
        - user
    SendMessageBody:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Activity Log
...